+++ app/src/main/java/com/br/matchmovies/view/ForgotPasswordActivity.kt
@@ -1,8 +1,8 @@
-package com.br.matchmovies
+package com.br.matchmovies.view
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import androidx.appcompat.widget.Toolbar
+import com.br.matchmovies.R
 
 class ForgotPasswordActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
