+++ app/src/main/res/layout/activity_forgot_password.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ForgotPasswordActivity">
+    tools:context=".view.ForgotPasswordActivity">
 
 
     <androidx.appcompat.widget.Toolbar
