+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".view.HomeActivity">
+
+
 
     <FrameLayout
         android:id="@+id/fl_wrapper"
@@ -18,7 +20,7 @@
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
         android:background="?android:attr/windowBackground"
-        app:itemBackground="@color/purple_700"
+        app:itemBackground="@color/purple_900"
         app:itemIconTint="#fff"
         app:itemTextColor="#fff"
         app:layout_constraintBottom_toBottomOf="parent"
