+++ app/src/main/res/layout/activity_main_login.xml
@@ -7,7 +7,7 @@
     android:layout_margin="16dp"
     android:layout_marginTop="10dp"
     android:orientation="vertical"
-    tools:context=".LoginActivity">
+    tools:context=".view.LoginActivity">
 
     <!--// login //-->
 
@@ -80,6 +80,10 @@
 
         </LinearLayout>
     </com.google.android.material.textfield.TextInputLayout>
+
+    <Space
+        android:layout_width="match_parent"
+        android:layout_height="20dp" />
 
     <TextView
         android:id="@+id/btesqueci"
