+++ app/src/main/res/layout/fragment_home.xml
@@ -10,7 +10,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".HomeActivity">
+        tools:context=".view.HomeActivity">
 
 
         <TextView
@@ -63,7 +63,7 @@
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:background="@color/purple_700"
+            android:background="@color/purple_900"
             android:foreground="@drawable/ic_logo_horizontal"
             android:foregroundGravity="center"
             app:layout_constraintTop_toTopOf="parent"
